fix(client): :bug: factorize send sms
